console: Move tabs to parent view
