+++ pkg/webui/views/console/device-overview/device-overview.styl
@@ -12,24 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-.id
-  h2()
-  nudge('up', 3px)
-  line-height: .8
-  height: 1rem
-  white-space: no-break
-  margin: 0
-  font-weight: 600
-  display: inline-block
+.overview-info
+  margin-top: $ls.s
 
-.head
-  margin-top: 0
-
-.title
-  margin: $ls.xxs 0
-  height: 2rem
-  display: flex
-  align-items: flex-end
+  &-general
+    border-normal('bottom')
+    box-sizing: border-box
+    padding-bottom: 2.55rem
 
 .table
   margin-top: $ls.xl
