+++ pkg/webui/views/console/device/device.styl
@@ -1,0 +1,36 @@
+// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+.title
+  margin: $ls.xxs 0
+  height: 2rem
+  display: flex
+  align-items: flex-end
+
+.head
+  margin-top: 0
+
+.id
+  h2()
+  nudge('up', 3px)
+  line-height: .8
+  height: 1rem
+  white-space: no-break
+  margin: 0
+  font-weight: 600
+  display: inline-block
+
+.tabs
+  border-normal('bottom')
+  margin-top: $ls.s
